Moving styling out of AudioPlayer
