+++ src/app/pages/LiveRadioPage/index.jsx
@@ -1,6 +1,12 @@
 import React, { useContext } from 'react';
 import { string, shape, object } from 'prop-types';
 import styled from 'styled-components';
+import {
+  GEL_SPACING,
+  GEL_SPACING_DBL,
+  GEL_SPACING_QUAD,
+} from '@bbc/gel-foundations/spacings';
+import { GEL_GROUP_2_SCREEN_WIDTH_MIN } from '@bbc/gel-foundations/breakpoints';
 import { Headline } from '@bbc/psammead-headings';
 import Paragraph from '@bbc/psammead-paragraph';
 import { useLocation } from 'react-router-dom';
@@ -19,6 +25,15 @@ import getEmbedUrl from '#lib/utilities/getEmbedUrl';
 const StyledGelPageGrid = styled(GelPageGrid)`
   width: 100%;
   flex-grow: 1; /* needed to ensure footer positions at bottom of viewport */
+`;
+
+const AudioPlayerWrapper = styled(AudioPlayer)`
+  width: calc(100% + ${GEL_SPACING_DBL});
+  margin: 0 -${GEL_SPACING};
+  @media (min-width: ${GEL_GROUP_2_SCREEN_WIDTH_MIN}) {
+    width: calc(100% + ${GEL_SPACING_QUAD});
+    margin: 0 -${GEL_SPACING_DBL};
+  }
 `;
 
 const LiveRadioPage = ({ pageData }) => {
@@ -107,7 +122,7 @@ const LiveRadioPage = ({ pageData }) => {
           <Paragraph script={script} service={service}>
             {bodySummary}
           </Paragraph>
-          <AudioPlayer
+          <AudioPlayerWrapper
             externalId={masterBrand}
             id={assetId}
             embedUrl={embedUrl}
